:recycle: [filestorage] Move disk storage tests
