+++ tests/unit/filestorage/adapters/test_disk.py
@@ -12,6 +12,7 @@ from cutty.filestorage.domain.files import Executable
 from cutty.filestorage.domain.files import File
 from cutty.filestorage.domain.files import RegularFile
 from cutty.filestorage.domain.files import SymbolicLink
+from cutty.filestorage.domain.storage import FileStorage
 from cutty.filesystems.domain.purepath import PurePath
 
 
@@ -207,3 +208,58 @@ def test_unknown_filetype(tmp_path: pathlib.Path) -> None:
     path = tmp_path.joinpath(*file.path.parts)
     assert not path.exists()
     assert not path.parent.exists()
+
+
+@pytest.fixture
+def storage(tmp_path: pathlib.Path) -> FileStorage:
+    """Fixture for a storage."""
+    return DiskFileStorage(tmp_path)
+
+
+def test_multiple_files(
+    tmp_path: pathlib.Path, storage: FileStorage, file: File, executable: File
+) -> None:
+    """It stores all the files."""
+    with storage:
+        storage.add(executable)
+        storage.add(file)
+
+    assert all(
+        tmp_path.joinpath(*each.path.parts).is_file() for each in (file, executable)
+    )
+
+
+def test_already_exists(storage: FileStorage, file: File) -> None:
+    """It raises an exception if the file already exists."""
+    with storage:
+        storage.add(file)
+        with pytest.raises(Exception):
+            storage.add(file)
+
+
+def test_overwrite_if_exists(tmp_path: pathlib.Path, file: File) -> None:
+    """It overwrites existing files."""
+    storage = DiskFileStorage(tmp_path, fileexists=FileExistsPolicy.OVERWRITE)
+
+    path = tmp_path.joinpath(*file.path.parts)
+    path.parent.mkdir()
+    path.write_text("old")
+
+    with storage:
+        storage.add(file)
+
+    assert path.read_text() != "old"
+
+
+def test_skip_if_file_exists(tmp_path: pathlib.Path, file: File) -> None:
+    """It skips existing files."""
+    storage = DiskFileStorage(tmp_path, fileexists=FileExistsPolicy.SKIP)
+
+    path = tmp_path.joinpath(*file.path.parts)
+    path.parent.mkdir()
+    path.write_text("old")
+
+    with storage:
+        storage.add(file)
+
+    assert path.read_text() == "old"
